Add logic to handle transposed letters
1. When we have an exact match, the decoder wasn't passing back the word in a
consistent way, so that's been added.
2. Add tests for a longer part of the first message. This has unintended
results, as the decode from the fragment

  I NEVER ONCE GABE YOU PREMISSIONS AND I DEMAND THAT TOU STOP MY ACCOUNT IS
  OREIVATE FOR A RWASOB BITCH

gets decoded to

  i never once (babe|gabs|gage|gave) you permissions and i demand that
  (toy|you) stop my account is irrigate for a reason bitch'

The decoder has provided four choices for GABE, one of which is correct --
that's fine.  It has found two choices for tou, one of which is OK. The mangled
word OREIVATE was expected to be handled later by deleting a letter and
arriving at 'private', but the algorithim has outsmarted me, and found
'irrigate' which is quite good -- but is three letters away from the original.
Hopefully, later versions will arrive at 'private', which is only two letters
away.