Add code to retry match by deleteing a letter
This was interesting, because it changed one of the correct words (premissions
-> permissions) to a shorter word that scored better (remissions) but makes
less sense. The result was:

  'i never once (babe|gabs|gage|gave|gab) you remissions and i demand that
  (toy|you|to) stop my account is private for a reason bitch'

So the good news is that it correctly found the word 'private' starting with
OREIVATE, but has lost permissions infavour of remissions.

Some code was copied and pasted, so there's some refactoring in the future of
this module.